refactor: move components closer to files
