+++ src/sections/dataset/dataset-files/files-table/file-info/file-info-cell/FileInfoCell.module.scss
@@ -30,7 +30,3 @@ display: flex;
   color: $dv-text-color;
   font-size: $dv-font-size-sm;
 }
-
-.labels-container >*{
-  margin-right: 0.5em;;
-}
