+++ src/sections/file/file-labels/FileLabels.module.scss
@@ -1,0 +1,3 @@
+.container > * {
+  margin-right: 0.5em;
+}
