+++ tests/component/sections/file/file-embargo/FileEmbargoDate.spec.tsx
@@ -1,6 +1,6 @@
-import { FileEmbargoDate } from '../../../../../../../../../src/sections/dataset/dataset-files/files-table/file-info/file-info-cell/file-info-data/FileEmbargoDate'
-import { FilePublishingStatus } from '../../../../../../../../../src/files/domain/models/FilePreview'
-import { FileEmbargoMother } from '../../../../../../../files/domain/models/FilePreviewMother'
+import { FileEmbargoDate } from '../../../../../src/sections/file/file-embargo/FileEmbargoDate'
+import { FilePublishingStatus } from '../../../../../src/files/domain/models/FilePreview'
+import { FileEmbargoMother } from '../../../files/domain/models/FilePreviewMother'
 
 describe('FileEmbargoDate', () => {
   it('renders the embargo date when embargo exists', () => {
